requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
asttokens==3.0.0
asyncio==3.4.3
attrs==24.3.0
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.5.0
cassandra-driver==3.29.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==2.0.0
colorama==0.4.6
comm==0.2.2
confluent-kafka==2.7.0
debugpy==1.8.11
decorator==5.1.1
distro==1.9.0
docopt==0.6.2
executing==2.1.0
faiss-cpu==1.9.0.post1
fastapi==0.115.6
ffmpy==0.5.0
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.12.0
geomet==0.2.1.post1
gradio==5.9.1
gradio_client==1.5.2
h11==0.14.0
hdfs==2.7.3
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.27.0
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.1
openai==1.58.1
orjson==3.10.12
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.1.1
pure_eval==0.2.3
py4j==0.10.9.7
pydantic==2.10.4
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.18.0
pyrsistent==0.20.0
pyspark==3.5.4
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
pywebhdfs==0.4.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
ruff==0.8.4
safehttpx==0.1.6
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
semantic-version==2.10.0
sentence-transformers==3.3.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.41.3
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.5.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websockets==14.1
yarl==1.18.3
